+++ src/001-Ramotion-Cinemood/RamotionCinemood.story.jsx
@@ -3,7 +3,7 @@ import { storiesOf } from '@storybook/react';
 import { action } from '@storybook/addon-actions';
 import RamotionCinemood from './RamotionCinemood.component';
 
-storiesOf('Ramotion: Cinemood Product Card Concept', module)
+storiesOf('001 - Ramotion: Cinemood Product Card Concept', module)
 	.add('with text', () => (
 		<RamotionCinemood/>
 	));
