--- src/Ramotion-Cinemood/_RamotionCinemood.styles.scss
@@ -1,14 +1,0 @@
-.wrapper {
-	border-radius: 20px;
-	height: 400px;
-	position: relative;
-	display: flex;
-	flex-direction: row;
-	background-image: url("https://images.unsplash.com/photo-1478001517127-fccc92f54906?dpr=2&auto=format&fit=crop&w=1500&h=1000&q=80&cs=tinysrgb&crop=");
-	background-repeat: no-repeat;
-	background-size: cover;
-	box-shadow: 0 19px 38px rgba(0, 0, 0, 0.30), 0 15px 12px rgba(0, 0, 0, 0.22);
-	margin: 20px;
-	font-family: "Roboto", sans-serif;
-	max-width: 900px;
-}
