+++ src/styles/_layout.scss
@@ -1,0 +1,72 @@
+@mixin flex-container-modifiers() {
+	&.flex-wrap {
+		flex-wrap: wrap;
+	}
+
+	&.space-between {
+		justify-content: space-between;
+	}
+
+	&.align-center {
+		align-items: center;
+	}
+
+	&.justify-center{
+		justify-content: center;
+	}
+
+	&.justify-end{
+		justify-content: flex-end;
+	}
+}
+
+//Flex Containers
+.row {
+	display: flex;
+	flex-direction: row;
+	flex-wrap: nowrap;
+	box-sizing: border-box;
+	// justify-content: flex-start;
+	// align-items: flex-start;
+	@include flex-container-modifiers;
+
+	@media screen and (max-width:600px){
+		width:100%;
+		flex-basis: 100%;
+		flex-wrap:wrap;
+	}
+}
+
+.column {
+	display: flex;
+	flex-direction: column;
+	flex-wrap: nowrap;
+	box-sizing: border-box;
+	// justify-content: flex-start;
+	// align-items: flex-start;
+	@include flex-container-modifiers;
+}
+
+.flex{
+	flex:1;
+	box-sizing: border-box;
+
+	&.half{
+		width:50%;
+		flex-basis:50%;
+
+		@media screen and (max-width:600px){
+			width:100%;
+			flex-basis: 100%;
+		}
+	}
+
+	&.full{
+		width:100%;
+		flex-basis:100%;
+	}
+}
+
+.row-wrap{
+	flex-wrap:wrap;
+}
