+++ src/styles/index.scss
@@ -1,3 +1,3 @@
 //@import 'https://fonts.googleapis.com/css?family=Roboto';
 
-@import '../Ramotion-Cinemood/RamotionCinemood.styles';
+@import '../001-Ramotion-Cinemood/RamotionCinemood.styles';
